Fix etc

+++ scripts/after_install.sh
@@ -7,5 +7,7 @@ chmod -R 777 $MAGE_ROOT
 echo "Restart nginx"
 sudo service nginx restart
 
-echo "Add link from efs to magento folder"
-ln -s /mnt/efs/etc/ /var/www/html/magento/app/
+if [ ! -f "${MAGE_ROOT}/app/etc" ]; then
+    echo "Add link from efs to magento folder"
+    ln -s /mnt/efs/etc/ /var/www/html/magento/app/
+fi
